relocated and cleaned up HackathonRepository
